+++ docs/.vitepress/generate-router.js
@@ -1,7 +1,31 @@
 const Router = {
-  Dom: ['useClickAway', "useClipboard", 'useDocumentVisibility', 'useElementBounding', 'useEventListener', 'useFullscreen', 'useHover', 'useRect', 'useResizeObserver', 'useTitle', 'useVirtualList', 'useWindowSize'],
-  State: ['useBoolean', 'useDebounce', 'useLocalStorage', 'useOnline', 'useSessionStorage', 'useThrottle', 'useToggle'],
+  Dom: [
+    'useClickAway',
+    'useClipboard',
+    'useDocumentVisibility',
+    'useElementBounding',
+    'useEventListener',
+    'useFullscreen',
+    'useHover',
+    'usePermission',
+    'useRect',
+    'useResizeObserver',
+    'useTitle',
+    'useVirtualList',
+    'useWindowSize',
+  ],
+  State: [
+    'useBoolean',
+    'useDebounce',
+    'useLocalStorage',
+    'useMounted',
+    'useOnline',
+    'useSessionStorage',
+    'useSupported',
+    'useThrottle',
+    'useToggle',
+  ],
   Worker: ['useCountDown'],
-  Effect: ['useDebounceFn', 'useThrottleFn']
+  Effect: ['useDebounceFn', 'useThrottleFn'],
 };
 export default Router;