+++ docs/useMounted/demo/index.vue
@@ -1,0 +1,8 @@
+<template>
+  <div>{{ isMounted ? 'mounted' : 'unmounted' }}</div>
+</template>
+
+<script setup>
+  import { useMounted } from 'zhongjiayao_v3_hooks';
+  const isMounted = useMounted();
+</script>